Remove error dumping proposal
the only part the user puts his effort is selections as mapping is
generated automatically. and selections can only be changed via
run_selection, which doest not dump its internal selection
representation on error. it is useless to dump selections without change
of the selector_view at this point.

we can also consider making the move/copy of files immediate as the user
makes his selection. this will simplify many parts (dumping selection
when error, pipelining options.. etc.) but would deprive the ability to
check or verify the decisions.